Move Polygon component to base layer folder.
