+++ frontend/src/features/map_planning/layers/base/components/polygon/PolygonTypes.ts
